requirements.txt
langchain-groq
python-dotenv
langchain-core
flask
flask-cors
